requirements.txt
python-telegram-bot==20.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
aiofiles==23.2.1
urllib3==2.0.7
validators==0.22.0
fake-useragent==1.4.0
